[DSpark] Support PP targets in aggregated serving and PD prefill with KV transfer
Rebuild of #56957 as a self-contained PR: #56956 was reverted by #58484 for
further streamlining, so this branch now carries the full stack on current
main:

- Reapply #56956 (DSpark IFB PP): PPHandler sampled-token/draft broadcast,
  deferred post-update warmup coverage, DSv4 router padding.
- DSpark prefill-only producer for PD: SpeculativeConfig.is_dspark_prefill_only
  + target_kv_transfer_config, NIXL guard, materialize_context_kv() called in
  place of propose() on the producer, drafter runs context_kv_only there.
- Merge with #57632 (context K/V precompute captured in the draft CUDA graph):
  the shared prepare half of propose()/materialize_context_kv() is factored
  into _prepare_draft_inputs(); propose() keeps the graph-aware store, the
  producer stores eagerly via _precompute_context_kv().
- Fix _post_update_kernel warmup to use the int32 idx_mapping serving builds.

+++ tests/v1/worker/test_pp_utils.py
@@ -160,7 +160,10 @@ def test_warmup_pp_decode_update_matches_serving_specialization(monkeypatch):
     idx_mapping, _, _, output_bin_counts = args[:4]
     sampled_tokens, num_sampled, num_rejected, query_start_loc = args[4:8]
     broadcast_drafts, draft_tokens_out = args[10:12]
-    assert idx_mapping.tolist() == [-1] and idx_mapping.dtype == torch.int64
+    # Serving always passes an int32 idx_mapping (both the unfiltered path and
+    # the freed-request filtered path); an int64 warmup tensor would compile a
+    # specialization that serving never reuses.
+    assert idx_mapping.tolist() == [-1] and idx_mapping.dtype == torch.int32
     assert output_bin_counts is None
     assert query_start_loc is None
     assert sampled_tokens.shape == (1, 3) and sampled_tokens.dtype == torch.int64
